requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.1
requests==2.32.3
geopy==2.4.1
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-dotenv==1.0.1